Adding testing files for cyberseceval,mlcommons and updated answercarefully
Motivation and Context
Testing the safety of responses in multiple languages ensures that LLMs remain inclusive, ethical, and reliable for a diverse global audience. It also helps identify and address biases, cultural differences, and technical gaps in training data, leading to better performance and user trust.
Type of Change

How to Test
We added the mlcommons 4 categories and updated cyberseceval datasets in the dataset folder. they are named cyberseceval2